flask_agri_backend/requirements.txt
flask==2.0.1
flask-restful==0.3.9
flasgger==0.9.7.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
python-dotenv==0.19.0
werkzeug==2.0.3
pymysql==1.1.1
sqlalchemy==1.4.52
redis==4.6.0